Update patch set 5
Patch Set 5:

(3 comments)

I like where this is going. As I understand it, all clients will support parsing 7 and 8 messages. When the server creates a message it can be 7 or 8 depending on the system property, defaulting to 7. Client-created messages are always version 7.

But it seems confusing that version 8 messages are sometimes JSON and sometimes not. It seems like if we decide to apply a hack we should set the version back to 7, so that version 7 messages are always non-JSON and version 8 messages are always JSON. That way we don't need a fallback when parsing; the version should always tell us what to do.

Patch-set: 5
Label: Verified=0