Fixes UiHandler method matching in generic classes
Method matching in UiHandler is completely broken for generic
ui fields as FieldWriterOfExistingType is using the raw type
of the ui field. Raw type of a generic field erases all generics
from the method signature. This was causing type parameter information
to be not used while matching methods. This is a huge problem with
some common events like SelectEvent and ValueChangeEvent.

We were not hitting the problem in some cases because JRawType is
broken and doesn't correctly delete the generics from the signature
of parent methods. However that was causing another problem:
UiBinder uses the type information for ui.xml so it doesn't really 
know the type parameters so it can't match generics methods from the
parent class (Issue 6091).

This fix changes the UiBinder generator so that
 - it will enhance the type information if there is a @uifield declaration
 - it will not use raw type in FieldWriterOfExistingType
 - it will fallback to erase type if it can't find any method match

Bugs: Issue 6091

Change-Id: I3121542b6eb4f06f36b88b02006b155422c45726