Update patch set 2
Patch Set 2:

Subpixel values was first introduced by IE10 I believe, so it is a rather new thing and I think not many have used them before.

The sub-pixel values comes handy when you need to do pixel perfect layouting which needs to work on different zoom levels as well. Widgets like tables and panels which need to measure the widgets and position them tightly beside each other really benefits from using the exact values rather than the rounded ones. We have a lot of those widgets in Vaadin. A common defect we usually see when not using sub pixel values is that on IE10 when user zooms in our widgets does not scale correctly and 1px gaps appear or then widgets start to wrap. It is in these cases subpixel would be preferred.

Basically it would be preferable for anyone to use the subpixel variants always when either measuring elements or doing pixel perfect positioning. Since we cannot change the current return values of the DOM methods we need a equally convenient way of using the subpixel values so people will prefer using them instead. If we hide them behind a cast people will still continue to use the round values and hit problems which we are seeing today.

Patch-set: 2