Update patch set 2
Patch Set 2:

Yeah I was afraid of that. If everyone should be using it, what then?

Splitting this CL does make sense; let's do that. However, punting the upgrade to 3.0 isn't a real solution for the subpixel problem. We still need to upgrade all the callers and, at least at Google, that can't be done by declaring a flag day and upgrading everyone at once.

So we would need two classes and this is looking like another Element upgrade, but even worse because we can't just make the old class a superclass of the new class like we did last time.

If we did have SubpixelElement then it's not pretty but it does provide a mechanism for people who care about this. There are a lot of occurrences of "Element" that would have to be changed. (I would want to call it "Elt" so at least we get a shorter code when we're done. :-)

If we decide to go through the trouble of a major upgrade then we should probably think about if we want to fix up and use Elemental.

Patch-set: 2
Reviewer: Gerrit User 5010 <5010@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0