Update patch set 5
Patch Set 5:

Hi Thomas,

the change is not about @RemoteServiceRelativePath. This is about making GWT RPC work when loaded from file system.

Because moduleBaseURL is final in RemoteServiceProxy there is no way of setting this at runtime today, it alway defaults to moduleBase and this is somewhat file:// for phonegap applications. (same for chrome extension)

While GWT RPC is really bad for applications where you are not in control of the update cycle it is still widely used with gwt-phonegap and there we use JSNI trickery to make it work. I would much rather see a straight forward way directly in GWT.

Patch-set: 5
Reviewer: Gerrit User 5385 <5385@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0