The job of defineSeed is to mark the existence of a class and associate its castMap with the class's various constructors.
FragmentExtractor chooses whether and what part of a defineSeed call to include in a fragment.

Previously this logic would only include a defineSeed function in a fragment if the fragment was known to make the particular type in question "live".
In this usage "live" means that if you have ALL other fragments loaded then you can not instantiate this type but the moment you additionally load
this fragment you become able to instantiate this type.

But it is possible that when you have ALL other fragments loaded you can already instantiate this type and so when you load this fragment you do not
expand the "liveness" of this type but you DO suddenly make certain of its constructors accessible that were not previously accessible. In this
situation previously logic would still not have included any defineSeed function at all and as a consequence these newly live constructors were never
getting their associated castMaps which led to runtime errors when casting an instance that had been created by one of these broken constructors.

This CL relaxes the defineSeed inclusion restriction to include fragments that are expanding a type's constructors liveness and not just fragments
that expand a type's liveness.

Change-Id: Idc87430b94338d289eae4760a1bef8ad127bd699