Update patch set 9
Patch Set 9:

Hi, we had a performance regression reported to us so we rolled this CL back internally.

The code where they noticed the regression is a builder class with many instance variables. Apparently the variable's default values were previously initialized on the prototype, so constructing the class was fast. After this change, they're initialized when the constructor is called.

In this particular builder it's pointless because it extends Object. So we could put in a check for that, or more generally we could search superclass constructors for virtual methods that might be overridden. If there aren't any, the difference isn't observable and we can initialize fields the faster way.

Or if that's too hard, conforming to the spec doesn't seem too important (given that we've lived without it for so many years), so we could just roll this back on trunk.

Patch-set: 9
Reviewer: Gerrit User 5010 <5010@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0