Update patch set 6
Patch Set 6: Code-Review+1

Nope, we can work with it - a wrapper isn't really an option, since we get real benefit from getting static methods that look like instance ones without the mess of static imports or new-ing up instances that wrap the original elt just long enough to run a method on them and gc them (unless the compiler now can unwrap that single final field and method and treat it as if it were static to begin with?). We used to have a great big flyweight thing going on to avoid the extra object creation, but move to extending Element to avoid that little extra cost...

My main point was to bring it up as a point to discuss, to make sure that adding a new final method to a non-final class was deliberate. GXT is happy to plan ahead and remove it before 2.6 lands (and thanks for the heads up on toggleClassName), and move to something else in the meantime, we'd just rather not make a habit of that, cutting new specific gxt minor releases to accommodate specific gwt minor releases, and I'm sure we're not alone in that.

Patch-set: 6
Reviewer: Gerrit User 5130 <5130@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+1