Update patch set 9
Patch Set 9:

Huh. I knew this correct way was "slower" but I'm surprised it actually showed up as a performance regression.

I agree looking for Object or super-cstrs with virtual calls is a good approach for keeping the old behavior. I can play with those.

I'm more curious about what we should do now in master--revert this commit there as well, with the idea that we can re-commit it when it has the optimization in place?

Alternatively, we could leave it in master, and make a separate CL to add the optimization, but then you guys will have to un-revert it for the optimization-only CL to apply cleanly.

Seems cleaner to revert it from master for now?

Patch-set: 9