Update patch set 6
Patch Set 6:

Thanks Goktug - one of the distinct advantages of extending Element is that we can subclass UiObject.getElement and already have access to the methods we want, and on the other side the subclass can invoke Element methods (without the pretend-jso-cast). As I understand it, from your example, GxtElement is superfluous - we'd need to .cast() to GxtElement to even get the .cast() method to get GxtExtensions ;).

(wrapping up, as Andrey's code review probably isn't the place...) W.R.T. devirtualization, that doesn't quite go far enough - constructors do not go away, and the this$static is still needed, even though it is only to grab the one field inside it. And while v8 might do the alias analysis and drop the ctor, I'm all by certain that IE6-9 don't (and 10 would be a guess).

Patch-set: 6