Update patch set 9
Patch Set 9:

Either way should be fine. We could have a visitor pass after all optimizations are done that marks hoistable fields. It does not need to be too precise, it might suffice just to check whether there is any call non static call in any of the constructors of the class + superclasses.

Patch-set: 9