Update patch set 9
Patch Set 9:

Oops, that wasn't very clear. Let me try again:

If think you're going to implement a patch this week to fix the performance regression, please go ahead and do it and don't worry about rollbacks. But if you think it's going to take a while, we should rollback this change on trunk so it doesn't get in the way of other compiler changes.

Patch-set: 9