Update patch set 2
Patch Set 2:

(4 comments)

An alternative would be to reintroduce a TreeLogger argument to the refresh() method (it was removed when the lazyResourceOracle was introduced), store it along with the needsRefresh flag, and later use it in doRefresh. And maybe also add a TreeLogger argument to getResourcesOracle.

That way, the log is in the branch for the code that called refresh(), rather than repeatedly appended to the branch for the code that created the ModuleDef.

Patch-set: 2
Label: Verified=0