Ensure clinits get called for JSO instance methods.
In following Java semantics, any static initialization code in a class
(collectively a clinit) should be run before any part of that class is used. To
achieve this, GWT ensures that this clinit was invoked before any (actually)
static method, and as part of any constructor. This way, instance methods are
sure to have already run the clinit, so it doesn't need to be run again.

However, JSO constructors don't exist, so instance methods are not guaranteed to
have had the clinit already run. This is the bug that this patch is trying to
fix.

Two locations are changed, both with the effect of ensuring that a clinit is
emitted for JSOs. First, the MethodInliner must not skip adding a call to the
clinit when the jso body is inlined and the method invocation (and ref to the
class) is dropped, and second, GenerateJavaScriptAST must likewise write out
explicit clinits for any static method or jso instance method.

Bug: issue 3192
Change-Id: If09885382bcf2d6c149cd8e48bfdd3ae527d67cb