Update patch set 1
Patch Set 1:

Afraid not, this is the exact same patch we looked at - only difference was that I pulled the sys.out.println I was using to debug the circumstances for this fix.

W.r.t. test cases, would those be best in the form of a test JSO that demonstrates the issue, or something that actually builds a sample ast and verifies that JsInvocations or JMethodCalls are created for the clinit? I have the test sample from the original bug report that could be made into a GwtTestCase, but that seems a little more testing the effect than the bug. That said, I'm not seeing (or perhaps not looking in the right place for) a lot of test cases that make sure the generated JS is sane.

One last thought - while moving this from SVN to Git, it occurred to me that this same fix could potentially be achieved in another way: In the same way that MakeCallsStatic ties newly static calls to their original instance version via JProgram.putStaticImpl, it appears that JsoDevirtualizer could do the same. This would make these methods show up as needing a clinit when the `program.isStaticImpl(x.getTarget())` check is made (as in MethodInliner.createClinitCall. From your general gwtc point of view, does that seem to be a reasonable (as well as clearer) fix?

Patch-set: 1
Label: Verified=0