Update patch set 9
Patch Set 9:

A good rule of thumb is; if you can't fix it in 5 min, you should revert. More time the trunk is left broken, more people are likely to get a bad copy on the sync; which is bad.

The cool thing is, Gerrit has a revert functionality; you just need to click a few buttons from the browser =)

Patch-set: 9