Update patch set 4
Patch Set 4:

So IE10 leaks as well without the workaround.

Sending out a repro to MS, but I think we have to deal with that. I would like to implement this without any major changes in other parts of the SDK, but we might want to open a bug just to be sure that we are not leaking memory on other occasions as well.

Patch-set: 4
Label: Verified=0