Update patch set 11
Patch Set 11: Code-Review-1

This patch does not work as it stands for a couple of reasons:

Special numbers (e.g. NaN) values are sent as strings, but they are in the string table which means that the receiver just gets an index to the table without knowing whether that is the actual value or just a string table index.

The sent value is not valid JSON in a couple of situations. This does at least include escaping unicode values using \x (JSON only supports \u) and sending long values as base64 strings enclosed using ' (JSON only supports ")

A readVersion method has been added to the super sourced version of ClientSerializationStreamReader but it is missing from the original version.

Patch-set: 11
Label: Code-Review=-1