Update patch set 5
Patch Set 5:

I would like to see this get landed or abandoned. I think having it would have a small value, but if Brian feels strongly I could also drop it.

At the very least we should update javadoc and all removeChildren to dom.Element 

@Hilbrand

A clear(true) will remove the history frame. But it is only used for IE6 anyway and we are phasing that out.

Patch-set: 5
Label: Verified=0