Update patch set 6
Patch Set 6:

Admittedly it's a judgment call. I'm somewhat biased in favor of shorter names and somewhat against giving two methods the same name. I don't think there's a sensible interpretation of clearing the DOM that doesn't include the widgets too, so it seems like we don't need to include that in the name.

(By the way, if we did have to do something tricky like avoid clearing the history frame then that would actually be an argument in favor of having this method.)

Patch-set: 6