Update patch set 7
Patch Set 7:

@Manuel good catch.

We do not need to care about the History Frame (I looked it up before submitting the patch), but we care about the iframe with code, especially when using Async Loading, because then we still need the frame.

I will follow up with a new patch addressing the bug issue 8200 I just filed.

Patch-set: 7