Update patch set 8
Patch Set 8:

The part that I'm not comfortable with is, you can exactly have the same effect with the current state. 

Yes you wouldn't have compiler support but I don't think you it is a big deal. Stub are going to be written once and you already did that. If a new method is added and if somebody calls it, if will fail-fast; and you can fix that yourself in your own code. With the interfaces, if someone forgets to add the new method to interface and then you will need a patch to gwt proper.

Being said that, rest assured, I wouldn't -1 for just not using it. My point was, what I suggested looked like pretty feasible and acceptable for your use case. If other reviewers +1's, you can go ahead submit the patch.

Patch-set: 8
Reviewer: Gerrit User 5110 <5110@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0