Update patch set 8
Patch Set 8:

What is the thinking for the remaining 10%-ish of widgets - all of the cell widgets (except CellPanel), remaining *boxes (DateBox, various number-Box), RichTextArea, missing *LayoutPanels (SplitLayoutPanel, LayoutPanel, HeaderPanel, DeckLayoutPanel), etc? Future patch, or less important? I could see skipping most of the boxes, since the api should be relatively similar since subclasses mostly modify internal behavior, but then you have IsNumberLabel which just makes the generics more explicit (and should almost certainly be extending IsValueLabel to get those other super-interfaces...). Then there are the layout panels which arguably look like HasWidgets.ForIsWidget, but if IsElement gets so specific as to allow direct dom manipulation, it seems reasonable to allow layout manipulation as well. 

AbstractHasData and subclasses (lots of cell widgets) is a nice abstract class which already has most -- but not all -- methods declared in various interfaces. 

Unless we're waiting for a future patch to finish the job, I'm not yet understanding the approach in which classes get selected for this treatment and which types/methods get a pass. I don't mean for this to stand in the way of this patch, but to suggest that the strategy is sufficiently unclear that we could probably use a gwt-site doc on this or a body of javadoc that covers how these are intended to be used, and how widget authors/maintainers should behave to make their code usable by downstream users who like these interfaces for testing.

Patch-set: 8
Reviewer: Gerrit User 5130 <5130@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0