Update patch set 2
Patch Set 2:

The point of using Java vs JSNI is that since java methods could be used in other parts, the usage of native functions like this.removeChild() would eventually be wrapped by the compiler with a function like rc().

In theory, this approach implies that native functions with long names would be replaced with short-name functions resulting a smaller code, noticeable if we call the method several times.

I have made a couple of tests to check what the compiler does, and I've seen that the compiler always is inlining calls to the java method removeChild() although it is used several times.

I dont know if the compiler is taking in account the number of times a method is used before considering whether is better to inline it, but aparently it is not doing so, I would ask to the mainling-list though.

- Java Implementation of removeAllChildren:
 public final void removeAllChildrenJava() {
   Node n;
   while((n = getLastChild()) != null)
     removeChild(n);
 }

- Application Java Code
 Window.alert("REMOVE-JAVA");
 Document.get().getBody().removeAllChildrenJava();
 Document.get().getBody().removeChild(null);

- Resulting Js code:
 Rg('REMOVE-JAVA');tb($doc.body);$doc.body.removeChild(null)}
 function tb(a){var b;while(b=a.lastChild)a.removeChild(b)};


- JSNI Implementation of removeAllChildren:
 public native final void removeAllChildrenJsni() /*-{
   while(this.lasChild) {
     this.removeChild(this.lasChild());
   }
 }-*/;

- Application Java Code
 Window.alert("REMOVE-JNSI");
 Document.get().getBody().removeAllChildrenJsni();
 Document.get().getBody().removeChild(null);

- Resulting Js code:
 Rg('REMOVE-JNSI');ub($doc.body);$doc.body.removeChild(null)}
 function ub(a){while(a.lasChild){a.removeChild(a.lasChild())}}


As you can see the code is almost the same in both cases, and in my tests they perform the same in prod (not in dev-mode though).

But the thing is that if gwtc were not inlined the functions the code should be a little smaller, something like this:
 function tb(a){var b;while(b=lc(a))rc(b)};

Anyway, this is just a curiosity, not a deal in this case.

Patch-set: 2