Update patch set 2
Patch Set 2:

Thanks for the detailed analysis. It is good to know that compiler can optimize the null checks.

For the inlining, AFAIK, it is pretty dumb; in practical terms, it just inlines one liners independent of how many places it is used.

If it followed what you have expected then the resulting code would probably be slower.

Patch-set: 2