Super Dev Mode: deemphasize unused Java lines in codeserver's UI
Modified the code server to render Java files as HTML when browsing
interactively. (JavaScript debuggers continue to load the plain text.)
In the HTML, lines that make no contribution to the JavaScript according
to the source map are greyed out.

This implementation is pretty inefficient since we parse the entire
source map on every html page load, but it seems to suffice.

Change-Id: Id16bcd2287630cebf7afb8bbbe0a198b28f5ada9