Update patch set 1
Patch Set 1: Code-Review-2

(3 comments)

Maybe add a comment that the other add() overloads are not overridden because only add(Widget) (and add(IsWidget), but it'll call add(Widget)) is common to RootPanel and RootLayoutPanel?

Or would you actually prefer to override all the add() overloads?

An alternative, which would also cover other misuses of layout panels, and which IMO is the true fix to the issue, would be to add a check to onAttach() of layout panels (where Layout#onAttach is called) testing whether the parent widget is a layout panel and/or whether the widget has been given an explicit size. The check would have to be scheduleFinally, or maybe even scheduleDeferred, to avoid false positives.

Patch-set: 1