Update patch set 1
Patch Set 1:

To be exact, scalar primitives are initialized using JType.getDefaultValue() that is overriden by JPrimitiveType to return a specific value for each primitive type (look at JPrimitiveType.java).

For arrays GWT distinguishes references that are left undefined, longs that are initialized to a LongLib type zero object, booleans that are initialized to false and all the rest of primitive types (char, float, double, int, etc?) that are initialized to the javascript value "0".

Patch-set: 1
Label: Verified=0