Update patch set 1
Patch Set 1:

I'm not familiar with Buck either, but if it's conceptually pretty similar to Google's build system (which it apparently tries to be), it might even be possible for us to write generic BUCK/BUILD files in a way to be compatible with both, and that would be excellent.

Out of curiosity, how do Maven and Buck interact?  Is it possible for us to use buck for building, and still benefit from Maven for third-party dependencies/etc?

Patch-set: 1
Reviewer: Gerrit User 5000 <5000@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0