Update patch set 1
Patch Set 1:

I'll answer both of you below, intertwined:

 > FYI, internally (not for GWT-SDK but for most of other
 > projects) we don't do monolithic BUILD files but instead
 > utilize package structures to divide project into
 > manageable parts.

As you can see (either here or in my mavenization work on gwt-dev), there are many cyclic dependencies between packages, so it wouldn't work so well without moving some files around anyway. As a quick throw-away hack, I didn't want to deal with many files and preferred one (or two) monolithic file.

 > I guess the build file was monolithic for GWT-SDK to
 > avoid adding BUILD files to source control.

I think the build became monolithic when ported to Ant years ago (maybe Blaze hadn't even already been invented) when GWT as open-sourced (with code.google.com being the source of truth) and never (re?)modularized when moved back to Google's infrastructure (possibly to avoid sprinkling BUILD files everywhere too).

 > We should really put this into consideration. We have a
 > lot of experience in our side to contribute to this and
 > make it better. It is easy to manage and also has pretty
 > good learning curve.

 > I'm not familiar with Buck either, but if it's
 > conceptually pretty similar to Google's build system
 > (which it apparently tries to be), it might even be
 > possible for us to write generic BUCK/BUILD files in a
 > way to be compatible with both, and that would be
 > excellent.

I have no preference between Buck or Maven for GWT (or whatever, as long as the build is modularized and maintainable), but the community is voicing concerns about using Buck to build GWT:

 * https://plus.google.com/113945685385052458154/posts/a4xiG76Q3Nm
 * https://plus.google.com/113945685385052458154/posts/9MPQYrbARok

AFAICT, the #1 reason we chose to move to Maven is its wide adoption, which lowers the barrier to entry for contributors.

See also http://www.javarants.com/2013/07/16/why-your-new-jvm-build-tool-is-making-things-worse-aka-use-maven-to-execute-your-build/

I'm also not sure how to run tests the way we need it (JRE tests, GWT tests in HtmlUnit, in real browsers, in dev mode or prod mode)

If you want to discuss this further, let's move to GWT-Contrib and/or the GWT G+ community, and/or one of the G+ posts linked above.

> Out of curiosity, how do Maven and Buck interact? Is it
> possible for us to use buck for building, and still
> benefit from Maven for third-party dependencies/etc?

Gerrit does that. There are a bunch of macros and genfile rules to download binaries from Central (rather than have them checked into a repository) and expose them as java_binaries, and then they're used as "standard" dependencies by java_library and java_test rules. The problem is mostly about transitive dependencies (Buck, like Blaze I suppose, doesn't make a difference between "build-only" and "needed at runtime" deps, it's an all or nothing switch), and how they change when you update a dependency.

But as I said in http://blog.ltgt.net/in-quest-of-the-ultimate-build-tool/, one could also use an external tool (such as Ivy, or something based on Aether or Maven) to "provision" the external dependencies (and even possibly generate a BUCK/BUILD file with all the transitive deps) before using Buck/Blaze to do the build and run the tests.

Patch-set: 1