Update patch set 1
Patch Set 1:

Moved from http://gwt-code-reviews.appspot.com/1357803/ and rebased on top of master

FYI, I need it for a project at work and reapplied it on top of 2.5.1 (not yet battle-tested).

In retrospect, I'm not sure it's a good move, I'll have to rethink about it later with a clear head.

Patch-set: 1
Label: Verified=0