Update patch set 1
Patch Set 1: Code-Review+1

For us, it's important that all systems select the same plural forms because they all go to the same translation UI where we use ICU4J to tell translators which number maps to which form (keyword). It would also be hard on QA if the behavior depended on the language.

It sounds like it would be ok to make "many" optional in this case, in GWT. If that produces warning messages, then there will be a lot of those...

Patch-set: 1
Reviewer: Gerrit User 5756 <5756@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+1