Update patch set 1
Patch Set 1:

(7 comments)

I'm a fan of the current behavior, actually. I think of GWT.log() as purely a developer's temporary debugging aid like adding print statements or setting breakpoints. Normally you won't check in code that has GWT.log() statements in it, but if you accidentally do they're harmless.

Regular logging is different since it's on in production and those logs are often sent back to the server. I tend not to use that for debugging because configurable logging is a source of confusion when debugging. GWT.log() is supposed to be easy to use and deterministic - if you call it, the message *will* be printed in Dev Mode (and now Super Dev Mode).

Patch-set: 1
Label: Verified=0