Update patch set 2
Patch Set 2:

So to be clear, we allow "many" to appear in translation files, but since GWT doesn't use them, they have no effect? Or is it possible for it to have an effect (if you turn it on)?

Based on the commit message it sounds like it's backward-compatible but there is an effect if messages include a "many" form.

Patch-set: 2