Update patch set 5
Patch Set 5:

I added a call to getEntityProxyType to check the transported type and avoid the need for ExtraTypes... Maybe better/more obvious to call validateTransportableType as Thomas suggests. I'm happy to change if requested.

I also added a RequestFactoryModelTest to check for a reasonable error message if the EntityProxyId is not parameterized.

There is no test coverage to check that the type is added to the set of proxies for the RequestContext. Would be pretty obvious if a regression in this area? More to the point I couldn't see any easy way to test this.

Patch-set: 5
Label: Verified=0