Update patch set 5
Patch Set 5: Code-Review+2

(1 comment)

If you want to test that it works even when the entity is only referenced by an EntityProxyId (I really wonder what would be the use-case, maybe when the entity is otherwise used in a distinct RequestContext?), then add another proxy (let's call it OnlyReferencedByEntityProxyId, or similar) for one of SimpleFoo or SimpleBar and add a corresponding service method.

I think it might be premature to add it if we don't have EntityProxy as a return type though, otherwise we'd only ever be able to send the stableId of a proxy that we just created on the client-side, and it'd be useless (except for checking how the server side deals with it). Unless we add a distinct RequestContext to get an instance. Might not be worth the hassle.

Patch-set: 5
Label: Code-Review=+2