Update patch set 5
Patch Set 5:

Thomas,

I think we are agreed then that more unit tests etc may be overkill at this time.

Even if there was a regression the error would be pretty blatant (denying knowledge of a Proxy that is clearing in the EntityProxyId).

Is the ability to return an EntityProxyId that useful? The EntityProxy (without with) is pretty light. If you really want to do it you could return the domain id (not type safe).

I may be having a senior moment but I don't currently have anything on my todo list of this patch.

Thanks for your patience.

Patch-set: 5