Update patch set 3
Patch Set 3:

[re: avoiding rebasing in export script]

So the current script takes an internal Perforce changelist, creates a new commit against HEAD, and then pushes to Gerrit.  There's no state that keeps track of what the last patch was against.

So it wouldn't be trivial to make it prefer to use the same parent commit as the last export, but I can add it if you think it would make reviewing easier.

Patch-set: 3