Update patch set 1
Patch Set 1:

That's true, and actually having that JavaScriptObject as second argument is at least unconfortable.
The patch was motivated by issue 6512 and also by the idea of being adherent to the getContext spec, that's not even true because of the varargs. 
May be better to eventually wait for a more meaningful second argument and go ahead with the factory workaround.

Patch-set: 1
Reviewer: Gerrit User 5705 <5705@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0