Update patch set 1
Patch Set 1:

Created a bug with more context:

https://code.google.com/p/google-web-toolkit/issues/detail?id=8289

I don't think this is the right fix, since the effect would be to silently treat the field as if it were transient if it's unused in GWT code. It seems better to either fix the bug or add a compiler error saying that the field should be marked transient.

(But since deRPC is deprecated I'm not sure it's worth fixing the right way.)

Patch-set: 1