Update patch set 1
Patch Set 1:

We could put a warning when this happens, and then the user can change the field to "transient" to suppress the warning. However, this code executes on the server so it might result in a lot of spam to the server logs. So I don't think we should emit a warning on every RPC call.

A compile-time error would be a better workflow, so it gets fixed before the release.

Patch-set: 1