Update patch set 1
Patch Set 1:

Another possibility might be to use ClientOracle.createUnusedIdent() to add extra fields to the object and read them back when the object is transmitted in the other direction.

However, this needs to be deterministic, since a different server in the cluster may do the deserialization.

Patch-set: 1
Label: Code-Review=0