Update patch set 1
Patch Set 1:

This is kind of urgent because it is blocking a new release. 

Changing the field to transient fixes the problem. Although I still have to check how many fields need to be changed and the possible side effects of it, this seems to be the best thing to do now.

Patch-set: 1