Update patch set 2
Patch Set 2:

(3 comments)

I would also add the handling of JDebuggerStatement in jjs.ToStringGenerationVisitor

It might also need to be handled in gflow/CfgBuilder.java unless we make sure it is gone if the dataflow pass is used. Probably an empty block visit will do (and omit the node from the graph which seems sensible), otherwise maybe a CfgSimpleNode needs to be added to the graph.

Patch-set: 2
Label: Verified=0